moved the App.js file
